backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
pydantic==2.7.1
pydantic_core==2.18.2
sqlparse==0.5.0
typing_extensions==4.11.0
